requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
aiosignal=1.3.1=pypi_0
antlr4-python3-runtime=4.9.3=pypi_0
anyio=4.4.0=pypi_0
appnope=0.1.4=pypi_0
argon2-cffi=23.1.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
arrow=1.3.0=pypi_0
asttokens=2.4.1=pypi_0
async-lru=2.0.4=pypi_0
attrs=23.2.0=pypi_0
babel=2.15.0=pypi_0
beautifulsoup4=4.12.3=pypi_0
bleach=6.1.0=pypi_0
bzip2=1.0.8=h80987f9_5
ca-certificates=2024.3.11=hca03da5_0
certifi=2024.2.2=pypi_0
cffi=1.16.0=pypi_0
charset-normalizer=3.3.2=pypi_0
circuitvizualization=1.0=dev_0
click=8.1.7=pypi_0
comm=0.2.2=pypi_0
contourpy=1.2.1=pypi_0
cycler=0.12.1=pypi_0
cython=0.29.37=pypi_0
debugpy=1.8.2=pypi_0
decorator=5.1.1=pypi_0
defusedxml=0.7.1=pypi_0
dsa=1.0.1=dev_0
exceptiongroup=1.2.1=pypi_0
executing=2.0.1=pypi_0
fastjsonschema=2.20.0=pypi_0
filelock=3.13.4=pypi_0
fonttools=4.51.0=pypi_0
fqdn=1.5.1=pypi_0
frozenlist=1.4.1=pypi_0
fsspec=2024.3.1=pypi_0
h11=0.14.0=pypi_0
hdbscan=0.8.33=pypi_0
httpcore=1.0.5=pypi_0
httpx=0.27.0=pypi_0
hydra-core=1.3.2=pypi_0
idna=3.7=pypi_0
ipykernel=6.29.4=pypi_0
ipython=8.25.0=pypi_0
ipywidgets=8.1.3=pypi_0
isoduration=20.11.0=pypi_0
jedi=0.19.1=pypi_0
jinja2=3.1.3=pypi_0
joblib=1.4.0=pypi_0
json5=0.9.25=pypi_0
jsonpointer=3.0.0=pypi_0
jsonschema=4.21.1=pypi_0
jsonschema-specifications=2023.12.1=pypi_0
jupyter=1.0.0=pypi_0
jupyter-client=8.6.2=pypi_0
jupyter-console=6.6.3=pypi_0
jupyter-core=5.7.2=pypi_0
jupyter-events=0.10.0=pypi_0
jupyter-lsp=2.2.5=pypi_0
jupyter-server=2.14.1=pypi_0
jupyter-server-terminals=0.5.3=pypi_0
jupyterlab=4.2.5=pypi_0
jupyterlab-pygments=0.3.0=pypi_0
jupyterlab-server=2.27.2=pypi_0
jupyterlab-widgets=3.0.11=pypi_0
kiwisolver=1.4.5=pypi_0
kneed=0.8.5=pypi_0
kooplearn=1.1.3=pypi_0
latent-circuit-inference=0.1=dev_0
libffi=3.4.4=hca03da5_0
llvmlite=0.42.0=pypi_0
markupsafe=2.1.5=pypi_0
matplotlib=3.4.3=pypi_0
matplotlib-inline=0.1.7=pypi_0
mistune=3.0.2=pypi_0
mpmath=1.3.0=pypi_0
msgpack=1.0.8=pypi_0
nbclient=0.10.0=pypi_0
nbconvert=7.16.4=pypi_0
nbformat=5.10.4=pypi_0
ncurses=6.4=h313beb8_0
nest-asyncio=1.6.0=pypi_0
networkx=3.3=pypi_0
notebook=7.2.1=pypi_0
notebook-shim=0.2.4=pypi_0
numba=0.59.1=pypi_0
numdifftools=0.9.41=pypi_0
numpy=1.26.4=pypi_0
omegaconf=2.3.0=pypi_0
openssl=3.0.13=h1a28f6b_0
overrides=7.7.0=pypi_0
packaging=24.0=pypi_0
pandas=2.2.2=pypi_0
pandocfilters=1.5.1=pypi_0
parso=0.8.4=pypi_0
pexpect=4.9.0=pypi_0
pillow=10.3.0=pypi_0
pip=23.3.1=py310hca03da5_0
platformdirs=4.2.2=pypi_0
pot=0.9.3=pypi_0
prometheus-client=0.20.0=pypi_0
prompt-toolkit=3.0.47=pypi_0
proplot=0.9.7=pypi_0
protobuf=5.26.1=pypi_0
psutil=6.0.0=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.2=pypi_0
pycparser=2.22=pypi_0
pygments=2.18.0=pypi_0
pynndescent=0.5.12=pypi_0
pyparsing=3.1.2=pypi_0
python=3.10.14=hb885b13_0
python-dateutil=2.9.0.post0=pypi_0
python-json-logger=2.0.7=pypi_0
pytz=2024.1=pypi_0
pyyaml=6.0.1=pypi_0
pyzmq=26.0.3=pypi_0
qtconsole=5.5.2=pypi_0
qtpy=2.4.1=pypi_0
ray=2.12.0=pypi_0
readline=8.2=h1a28f6b_0
referencing=0.35.0=pypi_0
requests=2.31.0=pypi_0
rfc3339-validator=0.1.4=pypi_0
rfc3986-validator=0.1.1=pypi_0
rnn-analysis=1.0=dev_0
rnn-coach=0.1=dev_0
rpds-py=0.18.0=pypi_0
scikit-learn=1.4.2=pypi_0
scipy=1.13.0=pypi_0
seaborn=0.13.2=pypi_0
send2trash=1.8.3=pypi_0
setuptools=68.2.2=py310hca03da5_0
six=1.16.0=pypi_0
sniffio=1.3.1=pypi_0
soupsieve=2.5=pypi_0
sqlite=3.41.2=h80987f9_0
stack-data=0.6.3=pypi_0
sympy=1.12=pypi_0
terminado=0.18.1=pypi_0
threadpoolctl=3.4.0=pypi_0
tinycss2=1.3.0=pypi_0
tk=8.6.12=hb8d0fd4_0
tomli=2.0.1=pypi_0
torch=2.3.0=pypi_0
tornado=6.4.1=pypi_0
tqdm=4.66.2=pypi_0
trainrnnbrain=1.0=dev_0
traitlets=5.14.3=pypi_0
types-python-dateutil=2.9.0.20240316=pypi_0
typing-extensions=4.11.0=pypi_0
tzdata=2024.1=pypi_0
umap=0.1.1=pypi_0
umap-learn=0.5.6=pypi_0
uri-template=1.3.0=pypi_0
urllib3=2.2.1=pypi_0
wand=0.6.13=pypi_0
wcwidth=0.2.13=pypi_0
webcolors=24.6.0=pypi_0
webencodings=0.5.1=pypi_0
websocket-client=1.8.0=pypi_0
wheel=0.41.2=py310hca03da5_0
widgetsnbextension=4.0.11=pypi_0
xz=5.4.6=h80987f9_0
zlib=1.2.13=h5a0b063_0
